Création des tests pour vérifier le parsing d'un deliveryplan

+++ src/optimodLyon/model/Intersection.java
@@ -20,7 +20,7 @@ public class Intersection
     /**
      * L'id de l'intersection
      */
-    private int id;
+    private long id;
 
     /**
      * Constructeur de la classe Intersection
@@ -28,7 +28,7 @@ public class Intersection
      * @param y La coordonnée y de l'intersection
      * @param id L'id de l'intersection
      */
-    public Intersection(float x, float y, int id)
+    public Intersection(float x, float y, long id)
     {
         this.x = x;
         this.y = y;
@@ -38,7 +38,7 @@ public class Intersection
     /**
      * @return L'id de l'intersection
      */
-    public int getId()
+    public long getId()
     {
         return this.id;
     }
